pyproject.toml
[project]
name = "newsagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gradio>=6.0.2",
    "ipykernel>=7.1.0",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-mistralai>=1.1.0",
    "langchain-openai>=1.1.0",
    "langgraph>=1.0.4",
    "mistralai>=1.9.11",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "typer>=0.20.0",
]
